Wang Tiles Draft
This is an initial draft of a corner-wang tiles for the 2d plane.
Why would we care about this?
1. Tiles.
2. Level generation.

The solver is very rough. The code quality is subpar and it isn't guaranteed
to produce a solution. In general I suspect that tiling the plane is NP-hard.
For our purposes, the right approach is to take an input tile set (that may
be hard), run the solver until it gets stuck, confabulate a tile that unsticks
the solver and restart. Once we've made a tiling easy, emit the augmented
tile set. Easier said than done, perhaps.

The other trick is to find a recursive subdivision of each tile. If we can do
that, tiling forever becomes trivial.