Break the Tile Generator
This should be a faithful implementation of the rowwise plane tiling algorithm
described in the Cohen paper. Unfortunately, it doesn't do the right thing. It
actually performs worse than the 'tile the plane randomly and then flip stuff
if you get stuck' algorithm. But, hey, let's check it in so someone can point
out how I'm being a dunce. The smug satisfaction of proving me wrong should be
enough to motivate someone to help me debug this.