Increase the value of sacrifices which remove existing muts (rchandra)
If you have a good mutation and get a conflicting bad mutation (e.g.
you have robust and gain frail) and that bad mutation is innate, you
lose all levels of the good mutation and gain the bad mutation.

Thus, Ru gives you an extra 15 piety for each such mutation lost unless
it's a stat boost, in which case and extra 6 piety.